+++ changelog.md
@@ -1,6 +1,10 @@
 
 
 # Changelog
+* 2021-01-29
+  * Added the `Polygon.signedArea` function (was global function in the demos before).
+  * Added the `Polygon.isClockwise` function.
+  * Added the `Polygon.area` function.
 * 2021-01-27 [v1.11.0]
   * Added the new svg renderer as integrated helper.
 * 2021-01-26